# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  model.AlertChannel:
    properties:
      channelType:
        description: '通道类型: email, webhook, wechat, dingtalk, feishu'
        type: string
      config:
        description: 通道配置（JSON格式存储）
        type: string
      createdAt:
        type: string
      enabled:
        description: 是否启用
        type: boolean
      id:
        type: integer
      name:
        description: 通道名称
        type: string
      updatedAt:
        type: string
    type: object
  model.AlertLog:
    properties:
      alertType:
        description: 告警类型
        type: string
      channelType:
        description: 发送通道
        type: string
      createdAt:
        type: string
      domain:
        description: 域名
        type: string
      domainMonitorId:
        description: 域名监控ID
        type: integer
      errorMsg:
        description: 错误信息
        type: string
      id:
        type: integer
      message:
        description: 告警消息内容
        type: string
      sentAt:
        description: 发送时间
        type: string
      status:
        description: '状态: success, failed'
        type: string
    type: object
  model.AlertReceiver:
    properties:
      createdAt:
        type: string
      dingtalkId:
        description: 钉钉ID
        type: string
      email:
        description: 邮箱地址
        type: string
      enableDingTalk:
        description: 接收钉钉
        type: boolean
      enableEmail:
        description: 告警方式偏好
        type: boolean
      enableFeishu:
        description: 接收飞书
        type: boolean
      enableSystemMsg:
        description: 接收系统消息
        type: boolean
      enableWeChat:
        description: 接收企业微信
        type: boolean
      enableWebhook:
        description: 接收Webhook
        type: boolean
      feishuId:
        description: 飞书ID
        type: string
      id:
        type: integer
      name:
        description: 接收人姓名
        type: string
      phone:
        description: 手机号
        type: string
      updatedAt:
        type: string
      userId:
        description: 关联的系统用户ID
        type: integer
      wechatId:
        description: 企业微信ID
        type: string
    type: object
  model.AlertReceiverChannel:
    properties:
      channelId:
        description: 通道ID
        type: integer
      config:
        description: |-
          在关联时使用的附加配置（JSON格式）
          用于存储通道特定的接收人配置，如飞书的user_id或phone等
        type: string
      createdAt:
        type: string
      id:
        type: integer
      receiverId:
        description: 接收人ID
        type: integer
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: 运维管理平台 API 文档
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: iom API
  version: "1.0"
paths:
...
  /monitor/alerts/channels:
    get:
      consumes:
      - application/json
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.AlertChannel'
            type: array
      summary: 获取告警通道列表
      tags:
      - AlertChannel
    post:
      consumes:
      - application/json
      parameters:
      - description: 告警通道信息
        in: body
        name: channel
        required: true
        schema:
          $ref: '#/definitions/model.AlertChannel'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.AlertChannel'
      summary: 创建告警通道
      tags:
      - AlertChannel
  /monitor/alerts/channels/{id}:
    delete:
      consumes:
      - application/json
      parameters:
      - description: 通道ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: success
          schema:
            type: string
      summary: 删除告警通道
      tags:
      - AlertChannel
    get:
      consumes:
      - application/json
      parameters:
      - description: 通道ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.AlertChannel'
      summary: 获取告警通道详情
      tags:
      - AlertChannel
    put:
      consumes:
      - application/json
      parameters:
      - description: 通道ID
        in: path
        name: id
        required: true
        type: integer
      - description: 告警通道信息
        in: body
        name: channel
        required: true
        schema:
          $ref: '#/definitions/model.AlertChannel'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.AlertChannel'
      summary: 更新告警通道
      tags:
      - AlertChannel
  /monitor/alerts/logs:
    get:
      consumes:
      - application/json
      parameters:
      - description: 页码
        in: query
        name: page
        type: integer
      - description: 每页数量
        in: query
        name: pageSize
        type: integer
      - description: 域名监控ID
        in: query
        name: domainMonitorId
        type: integer
      - description: 告警类型
        in: query
        name: alertType
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.AlertLog'
            type: array
      summary: 获取告警日志列表
      tags:
      - AlertLog
  /monitor/alerts/receiver-channels/:receiverId:
    get:
      consumes:
      - application/json
      parameters:
      - description: 接收人ID
        in: path
        name: receiverId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.AlertReceiverChannel'
            type: array
      summary: 获取接收人的通道关联列表
      tags:
      - AlertReceiverChannel
    post:
      consumes:
      - application/json
      parameters:
      - description: 接收人ID
        in: path
        name: receiverId
        required: true
        type: integer
      - description: 关联信息
        in: body
        name: req
        required: true
        schema:
          $ref: '#/definitions/model.AlertReceiverChannel'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.AlertReceiverChannel'
      summary: 添加接收人通道关联
      tags:
      - AlertReceiverChannel
  /monitor/alerts/receiver-channels/:receiverId/:channelId:
    delete:
      consumes:
      - application/json
      parameters:
      - description: 接收人ID
        in: path
        name: receiverId
        required: true
        type: integer
      - description: 通道ID
        in: path
        name: channelId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: success
          schema:
            type: string
      summary: 删除接收人通道关联
      tags:
      - AlertReceiverChannel
    put:
      consumes:
      - application/json
      parameters:
      - description: 接收人ID
        in: path
        name: receiverId
        required: true
        type: integer
      - description: 通道ID
        in: path
        name: channelId
        required: true
        type: integer
      - description: 配置
        in: body
        name: req
        required: true
        schema:
          additionalProperties:
            type: string
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.AlertReceiverChannel'
      summary: 更新接收人通道关联配置
      tags:
      - AlertReceiverChannel
  /monitor/alerts/receivers:
    get:
      consumes:
      - application/json
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.AlertReceiver'
            type: array
      summary: 获取告警接收人列表
      tags:
      - AlertReceiver
    post:
      consumes:
      - application/json
      parameters:
      - description: 告警接收人信息
        in: body
        name: receiver
        required: true
        schema:
          $ref: '#/definitions/model.AlertReceiver'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.AlertReceiver'
      summary: 创建告警接收人
      tags:
      - AlertReceiver
  /monitor/alerts/receivers/{id}:
    delete:
      consumes:
      - application/json
      parameters:
      - description: 接收人ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: success
          schema:
            type: string
      summary: 删除告警接收人
      tags:
      - AlertReceiver
    get:
      consumes:
      - application/json
      parameters:
      - description: 接收人ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.AlertReceiver'
      summary: 获取告警接收人详情
      tags:
      - AlertReceiver
    put:
      consumes:
      - application/json
      parameters:
      - description: 接收人ID
        in: path
        name: id
        required: true
        type: integer
      - description: 告警接收人信息
        in: body
        name: receiver
        required: true
        schema:
          $ref: '#/definitions/model.AlertReceiver'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.AlertReceiver'
      summary: 更新告警接收人
      tags:
      - AlertReceiver
...
securityDefinitions:
  Bearer:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"